Refactor: Blob Storage 처리 서비스 별도의 프로젝트로 분할
