+++ EveryPinApi.sln
@@ -19,6 +19,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EveryPinApi.Presentation", 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Shared", "Shared\Shared.csproj", "{DA7C6E93-07A1-4EDB-8744-3B17233BFA68}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExternalLibraryService", "ExternalLibraryService\ExternalLibraryService.csproj", "{A690419F-52C2-4E2B-B2C6-FA808ACAAE6A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -57,6 +59,10 @@ Global
 		{DA7C6E93-07A1-4EDB-8744-3B17233BFA68}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DA7C6E93-07A1-4EDB-8744-3B17233BFA68}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DA7C6E93-07A1-4EDB-8744-3B17233BFA68}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A690419F-52C2-4E2B-B2C6-FA808ACAAE6A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A690419F-52C2-4E2B-B2C6-FA808ACAAE6A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A690419F-52C2-4E2B-B2C6-FA808ACAAE6A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A690419F-52C2-4E2B-B2C6-FA808ACAAE6A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
